requirements.txt
fastapi
uvicorn
pydantic
ollama
requests
transformers
torch
accelerate
datasets
peft
bitsandbytes
pandas
openpyxl
yt-dlp
openai-whisper
librosa
soundfile
gradio
flask
flask-cors
